requirements.txt
base58
deepmerge
ed25519
Flask==0.12.4
Flask-AppBuilder==1.10.0
Flask-Babel==0.11.1
Flask-Caching==1.4.0
Flask-Compress==1.4.0
flask-googlemaps
Flask-Login==0.2.11
Flask-Migrate==2.2.1
Flask-OpenID==1.2.5
flask-qrcode
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.1
Flask-Testing==0.7.1
Flask-WTF==0.14.2
grpcio
grpcio-tools
pre-commit
pre-commit-hooks
pysha3
secp256k1
sphinx
sphinxcontrib-napoleon
toml
